+++ packages/mwp-api-proxy-plugin/src/proxy.test.js
@@ -9,8 +9,8 @@ import * as receive from './util/receive';
 
 import apiProxy$ from './proxy';
 
-jest.mock('mwp-cli/src/config', () => {
-	const config = require.requireActual('mwp-cli/src/config');
+jest.mock('mwp-config', () => {
+	const config = require.requireActual('mwp-config');
 	config.package = { agent: 'TEST_AGENT ' };
 	return config;
 });