+++ packages/mwp-config/localesSecondary.js
@@ -1,0 +1,4 @@
+module.exports = [
+	'en-AU',
+	'es-ES',
+];